Modify CSS for index
Set the body font-size to a value and also set the margin. H1 should be aligned to the center and margin should also be set.
The recipe-container card has to have a width, margin, padding and alignment. The links should act as block elements and the rest
should be set accordingly.